Update pl.json
My proposition of translate correction. "Funkcjonalnosci" is not good word for this tab & it's a common incorrect usage of this word by many Polish people. You can increase/decrease "funkcjonalnosc" but you cannot add/delete it so it never can be plural (unless we talk about two or more applications).